Add comprehensive tests for Set, Vector, and SortableVector implementations
- Implemented common set tests in set.hpp to compare SetLst and SetVec functionalities, including size, element existence, and min/max consistency.
- Added boundary value tests for sets to ensure proper handling of edge cases.
- Created order maintenance stress tests for sets to validate element ordering after various operations.
- Developed extended test functions for SetVec in setvec.hpp, covering insertion, binary search, predecessor/successor functionality, and removal operations.
- Introduced tests for Vector in vector.hpp, including constructors, copy/move semantics, resizing, access operations, and clear functionality.
- Added tests for SortableVector to verify sorting operations with random and already sorted data.

+++ projects/project_1/files/template/set/vec/setvec.cpp
@@ -169,7 +169,7 @@ namespace lasd {
         }
 
         long pos = FindSuccessorPos(data);
-        if (pos < 0 || pos >= size) {
+        if (pos < 0 || static_cast<ulong>(pos) >= size) {
             throw std::length_error("SetVec: Successor not found");
         }
 
@@ -183,7 +183,7 @@ namespace lasd {
         }
 
         long pos = FindSuccessorPos(data);
-        if (pos < 0 || pos >= size) {
+        if (pos < 0 || static_cast<ulong>(pos) >= size) {
             throw std::length_error("SetVec: Successor not found");
         }
 
@@ -199,7 +199,7 @@ namespace lasd {
         }
 
         long pos = FindSuccessorPos(data);
-        if (pos < 0 || pos >= size) {
+        if (pos < 0 || static_cast<ulong>(pos) >= size) {
             throw std::length_error("SetVec: Successor not found");
         }
 
@@ -214,7 +214,7 @@ namespace lasd {
         ulong pos = BinarySearch(data);
 
         // If the element is found, return false (no duplicate allowed)
-        if (pos < size && elements[pos] == data) {
+        if (static_cast<ulong>(pos) < size && elements[pos] == data) {
             return false;
         }
 
@@ -238,7 +238,7 @@ namespace lasd {
         ulong pos = BinarySearch(data);
 
         // If the element is found, return false (no duplicate allowed)
-        if (pos < size && elements[pos] == data) {
+        if (static_cast<ulong>(pos) < size && elements[pos] == data) {
             return false;
         }
 
@@ -262,7 +262,7 @@ namespace lasd {
         ulong pos = BinarySearch(data);
 
         // If the element is not found, return false
-        if (pos >= size || elements[pos] != data) {
+        if (static_cast<ulong>(pos) >= size || elements[pos] != data) {
             return false;
         }
 
@@ -281,7 +281,7 @@ namespace lasd {
 
         // Binary search to check if the element exists
         ulong pos = BinarySearch(data);
-        return (pos < size && elements[pos] == data);
+        return (static_cast<ulong>(pos) < size && elements[pos] == data);
     }
 
     template <typename Data>
@@ -326,7 +326,7 @@ namespace lasd {
         ulong pos = BinarySearch(data);
 
         // If the exact element is found
-        if (pos < size && elements[pos] == data) {
+        if (static_cast<ulong>(pos) < size && elements[pos] == data) {
             if (pos > 0) {
                 return pos - 1; // The predecessor is the element before
             }
@@ -352,8 +352,8 @@ namespace lasd {
         ulong pos = BinarySearch(data);
 
         // If the exact element is found
-        if (pos < size && elements[pos] == data) {
-            if (pos < size - 1) {
+        if (static_cast<ulong>(pos) < size && elements[pos] == data) {
+            if (static_cast<ulong>(pos) < size - 1) {
                 return pos + 1; // The successor is the element after
             }
             else {
@@ -362,7 +362,7 @@ namespace lasd {
         }
 
         // If the element is not found, the insertion point is the successor
-        if (pos < size) {
+        if (static_cast<ulong>(pos) < size) {
             return pos;
         }
